api/requirements.txt
fastapi
uvicorn
pydantic
redis
python-dotenv